move event listener to outside of function
removes tile event listener to outside of initialize turn function, as every time the function was called a new event listener was being added to the tile.
this resulted in click events causing functions to fire n+1 times. hillarious, but unnecessary. the event listener was moved to the play function, and upon clicking a tile the initializeTurn
function is called.