add conditional to check if tile is already taken and game is in play
Conditional prevents 'turn swapping' when clicking a tile that was already chosen by returning the function.
Previously, clicking on a tile that had already been selected would allow the next player (cpu if playing vs computer)
to take their turn, without the first player having played a piece at all. It also returns if the game
is not currently in play.